[src/nav_node.cpp] Standardization of topic name
The previous code syntax forced the launch file programmer to write
explicitly the name of each topic, even if those names will be fixed (eg
"action_topic", "nav_stop"...).
Moreover when you write "/" before the
name of the topic, the name is "absolute" and can't be affected by
namespaces.

I removed the parameters to choose the name of topics, if we want to do
that, we can use a "remap" as it is usually done.
I also changed the name of topics to allow the use of namespace

Signed-off-by: Thomas Bonnefille <thomas.bonnefille@outlook.com>