Connect nav_node obstacle gathering to cluster_detector
As it was observed last year, the kalmann filter used in lidar
triangulation introduce a small delay, this delay can become fatal for
this node.

In this commit I changed the obstacle detection to gather information
directly from the cluster detector, it imply that nav_node do not know
the exact number of obstacle on the playground nor the position in an
absolute coordinate system.

A new function has been introduce to convert circles from relative
coordinate to absolute coordinate.

Moreover, I unified the datatypes by considering the previous type
"Circle" as the new type "CircleObstacle" from cdf_msgs. This choice was
motivated by the fact that I had to rethink the obstacle processing
algorithm.

Signed-off-by: Thomas Bonnefille <thomas.bonnefille@outlook.com>